add proband relationship to bulk edit table
